requirements.txt
# 合同审查小助手 - 依赖包列表

# UI框架
PySide6==6.6.0

# 文档处理
python-docx==1.1.0
PyMuPDF==1.23.8

# AI/LLM框架
langchain==0.1.0
langchain-openai==0.0.2
langchain-community==0.0.10
openai==1.10.0

# 模板引擎
docxtpl==0.16.7
jinja2==3.1.3

# 工具库
python-dotenv==1.0.0
pydantic==2.5.3
requests==2.31.0

# PDF生成 (可选)
reportlab==4.0.7
